Split up the staking test
